Add --ignore-cpp option to ignore other cpp files
When --ignore-cpp is specified, the merger considers main-file
as the unique implementation file,
i.e. it assumes all header files included in main-file are
header-only libraries.

An example folder is presented to show the use case.
