+++ example/test-hello.cpp
@@ -1,0 +1,13 @@
+// this is an usused large cpp file,
+// or a file which needs not to be merged (e.g. file for unit test)
+
+void unusedFunction() {
+  /*
+
+
+
+
+
+
+  */
+}
