+++ example/world.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <string>
+std::string world() {
+  return "world";
+}
