+++ output.cc
@@ -1,0 +1,1 @@
+/*-- File: main.cc start --*/
